Revert "ion: remove ion_cp_heap"
This reverts commit 4700e767a5d2f3915ee6c6255883cade94e8953f.

+++ drivers/char/adsprpc.c
@@ -25,7 +25,6 @@
 #include <linux/hash.h>
 #include <linux/msm_ion.h>
 #include <mach/msm_smd.h>
-#include <mach/ion.h>
 #include <mach/iommu_domains.h>
 #include <linux/scatterlist.h>
 #include <linux/fs.h>
@@ -632,7 +631,7 @@ static int get_args(uint32_t kernel, struct smq_invoke_ctx *ctx,
 		}
 		list[i].num = 1;
 		pages[list[i].pgidx].addr =
-			buf_page_start((void *)(pbuf->phys +
+			buf_page_start((void *)(unsigned long)(pbuf->phys +
 						 (pbuf->size - rlen)));
 		pages[list[i].pgidx].size =
 			buf_page_size(pra[i].buf.len);
